# Laravel/projecte_final/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example for response examples value",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost/Projecte_Final/Laravel/projecte_final/public/index.php",
            "description": "L5 Swagger OpenApi Server"
        }
    ],
    "paths": {
...
        "/api/vehicles": {
            "get": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Veure totes els vehicles.",
                "operationId": "d8ca996263e988889652331c97b190f1",
                "responses": {
                    "200": {
                        "description": "Retorna totes les clients."
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/vehicle/{id}": {
            "get": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Veure totes els vehicles.",
                "operationId": "42b1174b588a58d6c525f62b951059a3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parameter with mutliple examples",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna totes les clients."
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Update vehicles",
                "operationId": "e395f36ab7fe92a03cea2ff0680b23c5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "matricula",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "1456LMH"
                    },
                    {
                        "name": "marca",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "BMW"
                    },
                    {
                        "name": "model",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Serie 1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Delete vehicle",
                "operationId": "0fe754763b180fb8ef61203ae034bcb8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Cita not found"
                    }
                }
            }
        },
        "/api/vehicle": {
            "post": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Adds a new vehicles",
                "operationId": "316598679acb121a4ac8d31b60e4528e",
                "parameters": [
                    {
                        "name": "matricula",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "1456LMH"
                    },
                    {
                        "name": "marca",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "BMW"
                    },
                    {
                        "name": "model",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Serie 1"
                    },
                    {
                        "name": "client_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        }
    }
}